pyproject.toml
[project]
name = "venture-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "fastapi>=0.135.3",
    "firecrawl-py>=4.22.1",
    "langchain>=0.2.0",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=0.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "plotly>=6.6.0",
    "pydantic>=2.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.56.0",
    "tabulate>=0.10.0",
    "uvicorn[standard]>=0.44.0",
]
